// types/achievements.type.ts
export enum Achievement {
    FIRST_WORKOUT = 'FIRST_WORKOUT',
    FIRST_LEVEL = 'FIRST_LEVEL',
    FIVE_WORKOUTS = 'FIVE_WORKOUTS',
    TEN_WORKOUTS = 'TEN_WORKOUTS',
    TWENTY_WORKOUTS = 'TWENTY_WORKOUTS',
    FIFTY_WORKOUTS = 'FIFTY_WORKOUTS',
    HUNDRED_WORKOUTS = 'HUNDRED_WORKOUTS',
    FIRST_PRESTIGE = 'FIRST_PRESTIGE',
    FIFTH_PRESTIGE = 'FIFTH_PRESTIGE',
    TENTH_PRESTIGE = 'TENTH_PRESTIGE',
}